moved tests with external dependencies to integration tests
